requirements.txt
flask>=3.0.0
openai-whisper>=20231117
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0
librosa>=0.10.0
